requirements.txt
Django>=4.2,<4.3
asgiref
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7